fix test

+++ tests/test_gemm.py
@@ -14,10 +14,10 @@ import extension_sycl
 A = torch.randn(32, 128, dtype=torch.bfloat16, device="xpu")
 B = torch.randn(128, 64, dtype=torch.bfloat16, device="xpu")
 
-ref_xpu = torch.matmul(A, B).to(torch.float)
+ref_xpu = torch.matmul(A, B)
 
 C = torch.zeros(32, 64, dtype=torch.float, device="xpu")
-actual_xpu = torch.ops.extension_sycl.cutlass_gemm(A, B, C)
+actual_xpu = torch.ops.extension_sycl.cutlass_gemm(A, B, C).to(torch.bfloat16)
 
 torch.testing.assert_close(ref_xpu.cpu(), actual_xpu.cpu())
